requirements.txt
# OpenAI Agents SDK
openai-agents>=0.7.0

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Environment management
python-dotenv>=1.0.0

# Data validation
pydantic>=2.6.0
